move getCategoricalTermFilter() from shared to client code; strip tw properties, that are not needed by the server, from the request payload
